feat(catalogue): establish shared read-only contract
Adopt one defensive catalogue implementation across the static Explorer and future transports, add bounded candidate schemas, and record the v0.2 tool lifecycle.

Refs #19
