+++ changelog.d/MCP-201.changed.md
@@ -1,0 +1,3 @@
+- Defined the `v0.2.0` read-only tool lifecycle and exact five-tool supported
+  target, and adopted a shared catalogue foundation in the existing static Explorer
+  without creating a service listener.
