+++ docs/decisions/README.md
@@ -12,6 +12,7 @@ research recommendations remain unchanged in the preserved pack at
 - [ADR-0006: Reviewed public geospatial examples](ADR-0006-reviewed-public-geospatial-examples.md)
 - [ADR-0007: Immutable GitHub Pages publication](ADR-0007-immutable-pages-publication.md)
 - [ADR-0008: Supported Pages transport from verified payload](ADR-0008-supported-pages-transport.md)
+- [ADR-0009: Read-only MCP tool lifecycle](ADR-0009-read-only-mcp-tool-lifecycle.md)
 
 Research decisions D01–D19 are accepted as constraints for building and evaluating
 Stage 0, not as blanket authority for later production stages. Research decision D20
