+++ package.json
@@ -12,7 +12,7 @@
   "scripts": {
     "build": "pnpm run build:okf && pnpm --recursive --if-present run build",
     "typecheck": "pnpm --recursive --if-present run typecheck",
-    "test:typescript": "pnpm --filter @gis-ai-go/mcp-gateway run test",
+    "test:typescript": "pnpm --filter @gis-ai-go/contracts run test && pnpm --filter @gis-ai-go/mcp-gateway run test",
     "test:explorer": "pnpm --filter @gis-ai-go/public-explorer run test:unit",
     "test:browser": "pnpm run build:explorer && pnpm --filter @gis-ai-go/public-explorer run test:browser",
     "test:python": "uv run --locked --cache-dir .uv-cache python -m unittest discover -s tests -p 'test_*.py' && uv run --locked --cache-dir .uv-cache python -m unittest discover -s services/geo-execution/tests -p 'test_*.py'",
