+++ packages/contracts/package.json
@@ -2,11 +2,18 @@
   "name": "@gis-ai-go/contracts",
   "version": "0.1.0",
   "private": true,
-  "description": "Shared GIS AI GO contracts and schema metadata",
+  "description": "Shared GIS AI GO contracts and governed catalogue semantics",
   "license": "MIT",
   "type": "module",
+  "exports": {
+    ".": {
+      "types": "./src/index.ts",
+      "import": "./dist/src/index.js"
+    }
+  },
   "scripts": {
     "build": "tsc --project tsconfig.json",
-    "typecheck": "tsc --project tsconfig.json --noEmit"
+    "typecheck": "tsc --project tsconfig.json --noEmit",
+    "test": "pnpm run build && node --test dist/test/*.test.js"
   }
 }
