+++ packages/contracts/src/catalogue/index.ts
@@ -1,0 +1,3 @@
+export * from "./catalogue.js";
+export * from "./links.js";
+export * from "./types.js";
