+++ packages/contracts/src/index.ts
@@ -1,5 +1,8 @@
 export const contractMetadata = Object.freeze({
   product: "GIS AI GO",
-  stage: 0,
+  contractVersion: 1,
+  lifecycle: "v0.2-foundation",
   schemaNamespace: "urn:gis-ai-go:schema",
 } as const);
+
+export * from "./catalogue/index.js";
