+++ schemas/catalogue-describe-request.schema.json
@@ -1,0 +1,34 @@
+{
+  "$schema": "https://json-schema.org/draft/2020-12/schema",
+  "$id": "urn:gis-ai-go:schema:catalogue-describe-request:v1",
+  "title": "GIS AI GO catalogue describe request",
+  "description": "A bounded request for one governed public catalogue record.",
+  "type": "object",
+  "additionalProperties": false,
+  "required": [
+    "record_id"
+  ],
+  "properties": {
+    "record_id": {
+      "type": "string",
+      "minLength": 1,
+      "maxLength": 512
+    },
+    "include": {
+      "type": "array",
+      "minItems": 1,
+      "maxItems": 2,
+      "uniqueItems": true,
+      "default": [
+        "relationships",
+        "sources"
+      ],
+      "items": {
+        "enum": [
+          "relationships",
+          "sources"
+        ]
+      }
+    }
+  }
+}
