+++ schemas/catalogue-problem.schema.json
@@ -1,0 +1,111 @@
+{
+  "$schema": "https://json-schema.org/draft/2020-12/schema",
+  "$id": "urn:gis-ai-go:schema:catalogue-problem:v1",
+  "title": "GIS AI GO catalogue problem",
+  "description": "A bounded RFC 9457-compatible problem envelope for catalogue requests.",
+  "type": "object",
+  "additionalProperties": false,
+  "required": [
+    "schema",
+    "type",
+    "title",
+    "status",
+    "code",
+    "request_id",
+    "trace_id"
+  ],
+  "properties": {
+    "schema": {
+      "const": "gis-ai-go.catalogue-problem.v1"
+    },
+    "type": {
+      "type": "string",
+      "format": "uri",
+      "maxLength": 2048
+    },
+    "title": {
+      "type": "string",
+      "minLength": 1,
+      "maxLength": 128
+    },
+    "status": {
+      "type": "integer",
+      "minimum": 400,
+      "maximum": 599
+    },
+    "code": {
+      "enum": [
+        "invalid_request",
+        "invalid_cursor",
+        "record_not_found",
+        "not_acceptable",
+        "rate_limited",
+        "complexity_limit_exceeded",
+        "service_unavailable",
+        "internal_error"
+      ]
+    },
+    "detail": {
+      "type": "string",
+      "minLength": 1,
+      "maxLength": 1024
+    },
+    "instance": {
+      "type": "string",
+      "format": "uri-reference",
+      "maxLength": 2048
+    },
+    "request_id": {
+      "type": "string",
+      "minLength": 1,
+      "maxLength": 128,
+      "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$"
+    },
+    "trace_id": {
+      "type": "string",
+      "pattern": "^[0-9a-f]{32}$"
+    },
+    "retry_after_seconds": {
+      "type": "integer",
+      "minimum": 1,
+      "maximum": 3600
+    },
+    "errors": {
+      "type": "array",
+      "minItems": 1,
+      "maxItems": 20,
+      "uniqueItems": true,
+      "items": {
+        "type": "object",
+        "additionalProperties": false,
+        "required": [
+          "path",
+          "code",
+          "message"
+        ],
+        "properties": {
+          "path": {
+            "type": "string",
+            "minLength": 1,
+            "maxLength": 256
+          },
+          "code": {
+            "enum": [
+              "required",
+              "invalid_type",
+              "invalid_value",
+              "out_of_range",
+              "not_unique",
+              "unknown_property"
+            ]
+          },
+          "message": {
+            "type": "string",
+            "minLength": 1,
+            "maxLength": 512
+          }
+        }
+      }
+    }
+  }
+}
